Added payoff calculations so server tracks intervals in a given strategy and updates payoffs accordingly. Added Countdown and Pointsdisplay to the client view, hooked them up to the correct variables, working well.
git-svn-id: file:///svn/CONGv2.0/trunk@1243 637b740e-b1db-41bf-a134-2c17f54ac410

+++ src/edu/ucsc/leeps/fire/cong/server/TwoStrategyPayoffFunction.java
@@ -6,7 +6,7 @@ package edu.ucsc.leeps.fire.cong.server;
  */
 public interface TwoStrategyPayoffFunction {
 
-    public int getPayoff(
+    public float getPayoff(
             float percent,
             float A, float B,
             float a, float b);
